refactor: split settings.rs
Split `settings.rs` into smaller files to make everything more
maintainable.
