+++ src/settings/keyless.rs
@@ -1,0 +1,18 @@
+use kubewarden::host_capabilities::verification::KeylessInfo;
+use serde::{Deserialize, Serialize};
+use std::collections::HashMap;
+use std::fmt;
+
+#[derive(Serialize, Deserialize, Debug)]
+#[serde(rename_all = "camelCase")]
+pub(crate) struct Keyless {
+    pub(crate) image: String,
+    pub(crate) keyless: Vec<KeylessInfo>,
+    pub(crate) annotations: Option<HashMap<String, String>>,
+}
+
+impl fmt::Display for Keyless {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "Keyless signature for image {}", self.image)
+    }
+}
