+++ src/settings/pub_keys.rs
@@ -1,0 +1,17 @@
+use serde::{Deserialize, Serialize};
+use std::collections::HashMap;
+use std::fmt;
+
+#[derive(Serialize, Deserialize, Debug)]
+#[serde(rename_all = "camelCase")]
+pub(crate) struct PubKeys {
+    pub(crate) image: String,
+    pub(crate) pub_keys: Vec<String>,
+    pub(crate) annotations: Option<HashMap<String, String>>,
+}
+
+impl fmt::Display for PubKeys {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "Pub key signature for image {}", self.image)
+    }
+}
